move verify_img to different file and create different implementation for Blender
